TIP-860: Move Category related api tests
