+++ tests/back/Pim/Enrichment/EndToEnd/Category/ExternalApi/ListCategoryEndToEnd.php
@@ -1,12 +1,11 @@
 <?php
 
-namespace Akeneo\Tool\Bundle\ApiBundle\tests\integration\Controller\Category;
+namespace AkeneoTest\Pim\Enrichment\EndToEnd\Category\ExternalApi;
 
-use Akeneo\Test\Integration\Configuration;
 use Akeneo\Tool\Bundle\ApiBundle\tests\integration\ApiTestCase;
 use Symfony\Component\HttpFoundation\Response;
 
-class ListCategoryIntegration extends ApiTestCase
+class ListCategoryEndToEnd extends ApiTestCase
 {
     public function testListCategories()
     {
